import: Script to import CSCP versions from the archives
`import` downloads the requested revision (if present), extracts the
files from the ZIP archive into a working directory, and then uses the
`git-commit-filetree` submodule to add those files as a new commit to
the tip of the local `vendor` branch.

There's some manual work that needs to be done to avoid errors; this
is documented in the README, where we also expanded the end-user
documentation a bit.